+++ R/Interpolate.R
@@ -7,6 +7,9 @@
 #' @param x.out,y.out x and y values where to interpolate (see Details)
 #' @param data optional data.frame with the data
 #' @param grid logical indicating if x.out and y.out define a regular grid.
+#' @param path a logical or character indicating if the x.out and y.out define a
+#' path. If character, it will be the name of the column returing the order of
+#' said path.
 #'
 #' @details
 #' `formula` must be of the form VAR1 | VAR2 ~ X + Y where VAR1, VAR2, etc...
@@ -17,6 +20,13 @@
 #'
 #' If `grid = TRUE`, `x.out` and `y.out` must define the values of a regular
 #' grid. If `grid = FALSE`, they define the locations where to interpolate.
+#' Both `grid` and `path` cannot be set to `TRUE` and the value of `path` takes
+#' precedence.
+#'
+#' `x.out` can be a list, in which case, the first two elements will be interpreted
+#' as the x and y values where to interpolate and it can also have a `path` element
+#' that will be used in place of the `path` argument. This helps when creating a
+#' path with [as.path] (see Examples)
 #'
 #' @return
 #' A data.frame with interpolated values and locations
@@ -40,10 +50,33 @@
 #' geopotential[, c("u", "v") := GeostrophicWind(gh, lon, lat)]
 #' interpolated <- geopotential[, Interpolate(u | v ~ lon + lat, x.out, y.out)]
 #'
+#' # Interpolate values following a path
+#' lats <- c(-34, -54, -30)   # start and end latitudes
+#' lons <- c(302, 290, 180)   # start and end longituded
+#' path <- geopotential[, Interpolate(gh ~ lon + lat, as.path(lons, lats))]
+#'
 #' @export
 #' @importFrom fields interp.surface
 #' @import data.table Formula formula.tools
-Interpolate <- function(formula, x.out, y.out, data = NULL, grid = TRUE) {
+Interpolate <- function(formula, x.out, y.out, data = NULL, grid = TRUE, path = FALSE) {
+    if (is.list(x.out) == TRUE) {
+        path <- x.out$path %||% path
+        y.out <- x.out[[2]]
+        x.out <- x.out[[1]]
+    }
+
+    index <- NULL
+    if (!isFALSE(path) & !is.null(path)) {
+        if (isTRUE(path)) {
+            path <- ".order"
+        } else if (!is.character(path)) {
+            stop("`order` must be logical or character")
+        }
+        grid <- FALSE
+
+        index <- seq_along(x.out)
+    }
+
     dep.names <- formula.tools::lhs.vars(formula)
     if (length(dep.names) == 0) stop("LHS of formula must have at least one variable")
 
@@ -55,6 +88,9 @@ Interpolate <- function(formula, x.out, y.out, data = NULL, grid = TRUE) {
     formula <- Formula::as.Formula(formula)
     data <- as.data.table(eval(quote(model.frame(formula, data = data,
                                                  na.action = NULL))))
+    if (!.is.reggrid(data, ind.names)) {
+        stop("Interpolate need a unique value for each x and y")
+    }
 
     # Accomodate 1D interpolation
     if (length(ind.names) == 1) {
@@ -87,11 +123,16 @@ Interpolate <- function(formula, x.out, y.out, data = NULL, grid = TRUE) {
                  If x.out and y.out define unique points on a regular grid, use grid = TRUE')
         }
         loc <- data.table(x.out, y.out)
+        if (!is.null(path) & !isFALSE(path)) {
+            set(loc, NULL, path, index)
+        }
     } else {
         stop('wrong mode, choose either "grid" or "locations"')
     }
-    colnames(loc) <- ind.names
+
+    colnames(loc)[seq_along(ind.names)] <- ind.names
     alloc.col(loc, ncol(data))
+
     dcast.formula <- as.formula(paste0(ind.names, collapse = " ~ "))
 
     for (v in seq_along(dep.names)) {
@@ -107,7 +148,52 @@ Interpolate <- function(formula, x.out, y.out, data = NULL, grid = TRUE) {
 }
 
 
+#' Interpolates between locations
+#'
+#' This is a helper function to quickly make an interpolated list of locations
+#' between a number of locations
+#'
+#' @param x,y numeric vectos of x and y locations. If one of them is of length 1,
+#' if will be reciceld.
+#' @param path either `TRUE` of a character vector with the name of the path.
+#'
+#' @return
+#'
+#' A list of components `x` and `y` with the list of locations and the `path`
+#' arguments
+#'
+#' @details
+#'
+#' This function is mostly useful when combined with [Interpolate]
+#'
+#' @seealso Interpolate
+#'
+#' @export
+as.path <- function(x, y, n = 10, path = TRUE) {
+    if (length(x) < 2 & length(y) < 2) {
+        stop("either `xs` or `ys` must be of length greater than 1")
+    }
 
+    if (!is.numeric(x)) {
+        stop("`x` must be a numeric vector")
+    }
 
+    if (!is.numeric(y)) {
+        stop("`y` must be a numeric vector")
+    }
 
+    if (length(x) == 1) {
+        x <- rep(x, length(y))
+    } else if (length(y) == 1) {
+        y <- rep(y, length(x))
+    }
 
+    points <- seq_len(length(x))
+    index <- seq(min(points), max(points), length.out = n)
+    index <- sort(unique(c(index, points)))
+    xs_interpolate <- approx(points, x, index)$y
+    ys_interpolate <- approx(points, y, index)$y
+
+    list(x = xs_interpolate, y = ys_interpolate,
+                path = path)
+}